+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Panels.Header` to match title style from the GUI in `sandstone/WizardPanels`
+
 ## [2.9.8] - 2025-02-04
 
 ### Fixed